Add extra_data param to `set_cfg_constructor` function
Summary:
Add an optional `extra_data` param to `set_cfg_constructor` function to add custom starlark data that is only suitable for use internally and not elsewhere. I want to attach some internal-only validation here in the follow-up diffs.

The alternative to this is to support completely separate cfg constructor functions between OSS and internal. I'd prefer not doing that since that would make testing a lot difficult. Instead, just support an optional param that is always going to be None in OSS and may be set to some value internally.

Reviewed By: samkevich

Differential Revision: D61836785

fbshipit-source-id: 7c43c24c5320bec2b11cd2062535ad5a5d833670