Make TD work with check_dependencies_test
Summary:
This was missed from the original implementation of `check_dependencies_test`. In order for the test to be triggered by CI when the target it's checking against (or its deps) are changed, we need `check_dependencies_test` to take an unconfigured dep on the test target. Without it, TD won't know to schedule this test.

This diff fixes this problem by adding the test target as a dep in the unconfigured graph but not in the configured graph.

Reviewed By: JakobDegen

Differential Revision: D61982286

fbshipit-source-id: 7d8530c9ac0f47ebdd7349f315b8593c49c883ee