Merge all index stores to local folder
Summary:
# This diff stack
This diff stack is to  prepopulate the index store for focus.
It will:
1. enable building index in build rule for swift and objective-c.
2. build the index store by given "index-store"/"all-index-store" sub target

 ---
## Context
Right now, the final merge action need to wait all the index stores to be materialized  and  then do the merge. It is not efficient. Since for our case we do not need wait for all index stores to merge

## This  diff

1. Read the config to get the local path
2. In merge_index_store.sh, use a environment variable flag to run the local merge action. It copies all the files to a local directory using a local only action
3. To make sure buck2 will run the actions, it need a dummy output for each action and create  another final merge output dummy output based on these dummy outputs
 ---

**Before change:**
 {F1676028598}
materialization stage: local will do nothing but wait. network bound. CPU do nothing.
merge stage: all are merging, cpu and i/o bound.
**With local merge**
 {F1676029792}
Whenever a index-store is materialized, it can merge to the final output. It can make the more use of computer resources: cpu and network.

In this way, the final merge action do not need to wait all the index stores to be materialized. Each local merge action will run immediately after the relevant index store materialized. It will speed up the process.

Reviewed By: milend

Differential Revision: D58143446

fbshipit-source-id: b256a607c526733439263381c319c2f5bfbf9c0c