Replace crash segault command with abort
Summary:
`buck2 debug seg-fault`/`buck2 debug crash segfault` is intended to test logging when buck crashes due to SIGSEGV, however it's broken because it hits a panic (`async fn resumed after completion` - P1556461758) either instead of or before the segfault is thrown or handled.

I couldn't get it to segfault from rust code without seeing that panic, and it's not a panic we see anywhere else.

Triggering a segfault should be mostly the same as sending SIGABRT as far as our logging is concerned, so let's just do that instead.

This is mostly useful to test behavior by C/C++ failures, esp. RE segfaults.

Reviewed By: stepancheg

Differential Revision: D61955974

fbshipit-source-id: 011c42f0cc48a9dfc02be950f523a5fbdd45553b