Add support for coerced selector
Summary:
# Context
We miss the support for selector and concat in CoercedAttr. So that in the unified node's attribute api, we do not need get attribute type anymore.

# This diff
This diff adds support for coerced selector.
It uses the symbol name `bxl.Selector` for it. (Don't know if it is a good name here.)

Reviewed By: stepancheg

Differential Revision: D61194383

fbshipit-source-id: 3ea9bd1c796491ec679637c39acfe6b4f2a89121