rust-project: Simplify scuba logging
Summary: Using tracing for this is neat, but too complex when we just need to log to scuba once at the end of the program. The stringly-typed interface between all of the logging calls and the final sample is also a bit uncomfortable. Manually logging the stuff we care about should be easier to understand and maintain.

Reviewed By: Wilfred

Differential Revision: D61995962

fbshipit-source-id: 9cb257668b405bf6d9ed463b900697efd999222c