rust-project: Include current revision in scuba logs
Summary:
Include the output of `sl id` in our scuba logs for `develop`. This could be useful in UPND investigations and it will let us measure how many people are on warm checkouts for Glean.

Note that `sl id` will tell us not only the revision but also include `+` if the checkout is dirty.

Reviewed By: davidbarsky

Differential Revision: D61998335

fbshipit-source-id: 37a85047fc153ad63ae46b943a4c37047bdd337b