Enable close_stdin on more targets
Summary:
When running a windows command genrule and interrupting it with CTRL+C, the process will be signalled, but the trampoline will prompt to `Terminate batch script (Y/N)?`. Failing to respond with `y` will leave processes hanging and mess up terminal input. The recently added `close_stdin` label allows us to opt-out of having to respond to this prompt, but this workaround only works if the trampoline batch file is calling another batch file.

This change generalizes this workaround so that it does not matter what the underlying executable/script is. We enable this by having the trampoline script call itself with a `<NUL` redirect, using an environment variable so we only recurse once.

Reviewed By: KapJI

Differential Revision: D62153022

fbshipit-source-id: 1326b3468a6221f994e3ba9780415609aeb96e79