Add stderr stack trace key to category key
Summary:
It's useful to be able to group / categorize crashes in the daemon in order to be able to automatically detect new crashes. These can't be automatically categorized with tags and it's painful to categorize these by string matching.

An example category key with a trace key looks like:
`buck2_client_ctx/src/daemon/client.rs:CLIENT_GRPC:SERVER_PANICKED:crash(bf6eeb)`

A lot of the logic here is to sanitize/trim the stack traces so that we don't get new hashes for every minor code change.

It might be ideal to just hash the individual function name that triggered the crash, but it's not guaranteed we can separate that from low level signal handling logic in the stack trace, so this hashes the function names for 5 lines in the stack trace and tries to exclude known platform/signal handling function names.

Reviewed By: ezgicicek

Differential Revision: D61807198

fbshipit-source-id: 66f60b3b096cc50ba50e895e03cc98b65c57a5a6