Make build count errors soft_errors
Summary:
We see about ~70 build failures due to this per day, possibly due to FS issues in CI. We should just pass the build and log a soft_error if it gets this far.

Previous diff will cause this to throw an error if you rebuild with different buck versions on the same revision, because it can't parse the json anymore. This isn't likely to affect users but also shouldn't fail the build.

Reviewed By: iguridi

Differential Revision: D60065077

fbshipit-source-id: 3fce0bfaf5484021d74e0c0c1a794030cb5d9df7