Enabling test bundling when running mac tests on RE
Summary:
In the context of migrating jobs running on mac-arm to Linux, when tests are sent to be run on a mac re worker, each test withing a test suite will represent one RE action, and the output of the action will reporesent one write to CAS storage

So if there exists 40 different tests in a suite, there will be 40 actions and the output of each action will be written individually to CAS, and this can have greatly increase the write QPS from the RE workers leading to throttling.

Bundling tests together will cause all 40 individual tests to grouped into one action for execution. Hence resulting in one write to CAS. This reduces the overall amound of data written to storage and SHOULD prevent use case throttling from remote execution

Reviewed By: akrieger

Differential Revision: D61924673

fbshipit-source-id: 0df8b89e6a7aac8005efb3e8f11e4146532abd68