Never use RE for test worker actions
Summary:
Test workers aren't configured correctly to work in RE at the moment. Buck might try to run them remotely but it will always fail and fall back to local in case of hybrid execution. This is bad because hybrid will borrow its command execution request into both executors. When only running locally, the request is consumed by the local executor and doesn't take up memory while the command is running.

This is important because Test Runway runs 2,000,000 large test jobs using a worker tool "locally." (In practice a worker distributes them across a bunch of sandcastle machines.) These jobs are super memory constrained partly because Buck keeps around the memory for each local action instead of dropping it. This is partly because the command is borrowed here (see lines 220-221) instead of owned. My understanding is that workers imply local execution anyway, and that the remote variant of this will fail. If that is the case, avoid the extra memory caused by borrowing, and just run a local action when a worker is requested.

Reviewed By: christolliday

Differential Revision: D62099575

fbshipit-source-id: cc586beefbeb3147729315b0f3b0108f85ff3ada