Make dep_only_incompatible soft errors not quiet
Summary: We've now cleaned up the major dep_only_incompatible targets and now are looking at rolling this out to be a hard error. As a first step, make this soft error no longer quiet so it's easier for people doing cleanup work to see this error from builds locally or from CI without digging into the logs.

Reviewed By: brchien

Differential Revision: D62277998

fbshipit-source-id: d8346e43a724979e12be67d60c27b452bd2a4f5f