Enable bundled prelude and modifier in configured_target bxl test
Summary: I need test adding `modifiers` in `configured_target` function in BXL in the next diff, so I need to enable modifiers in the isolated test case. The best way to do that today is via bundled prelude, so this diff is to enable bundled prelude and modifiers in `configured_target` test data this week.

Reviewed By: Nero5023

Differential Revision: D62112570

fbshipit-source-id: a7c26ce0d6d135c71d290b8d998d4768aec796bc