Clean-up generation of start_shell.sh
Summary:
# Context
The `erlang_shell.build_run_info()` function is used to build the executable associated to `buck2 run` on both `erlang_app()` and `erlang_test()` targets. For each target, it writes a `start_shell.sh` script that invokes erl with all necessary parameters.

For readibility, it puts each argument to erl on a separate line, so a `\` needs to be appended so that it is interpreted as a single command

The function accepts a `additional_args` arguments, currently used for passing extra arguments to the test-shell on `erlang_test()` targets.

# Problem
It is currently the responsibility of the caller to include necessary `"\"` delimiters inside `additional_args`. If they are omitted, things break as the command will end early. This is confusing, error-prone, and arguably leaks implementation details.

# This diff
We move the resnponsibility of adding the necessary `\` separators the the `build_run_info()` function itself.

Reviewed By: TheGeorge, acw224

Differential Revision: D62450307

fbshipit-source-id: 65736780c7cfc49167cb8e330246a4a3c4169fc7