host_cpu_load_info on macos
Summary:
`host_statistics(HOST_CPU_LOAD_INFO)`.

This function is not documented anywhere, written by digging stackoverflow and other websites.

[sysinfo crate is banned](https://fb.workplace.com/groups/buck2core/posts/828710522656019), so write it from scrach.

Context: to add this information to Scuba to understand when buck2 is slow because operating system is not doing anything.

It is not crystal clear whether this information is helpful to us.

In theory we can compute all CPU usage of buck2 and spawned processes, and if it is far from 100%, buck2 is not running optimally.

Or we can subtract system CPU usage from buck2 and children from system CPU usage, and if it is high, it means other processes are eating CPU.

Anyway, this implementation allows experimentation.

Reviewed By: ezgicicek

Differential Revision: D62207258

fbshipit-source-id: 25e2df912e89f386d15036da638cb82cfebc5ffe