add network type net io stat
Summary:
We can determine network type on windows using WinAPI. On Mac and Linux it's a bit more complicated. For now we just determine it based on name. en0 for  mac is always WiFi.
There is an API for MacOs `SCNetworkInterfaceGetInterfaceType` but we need to add dependency on Appple system-configuration framework, I'm going to do it later.

Reviewed By: ezgicicek

Differential Revision: D62440864

fbshipit-source-id: 975e4e9026b29af6b49bdbbce5b9a7a600953989