Move test_info functions to own module
Summary:
# Context
When running a erlang tesst, a json file with "test info" is written to disk and passed
to the test runner. We parse this file before running the test, to get info such as
the toolchain to use, providers, hooks, etc.

# Problem
At the moment the test-shell is not aware of this test-info file, so in particular it can't
the correct toolchain, etc. We want the test-shell to load this file as well, but then we
want to have the "parsing" logic in one place, so it can be reused

# This diff
We move the loading of the test info file to its own module. It gets used by the
test-shell later in the stack.

Reviewed By: TheGeorge

Differential Revision: D62500827

fbshipit-source-id: e720fd659eac20a4b7bcc34c22259c243c188aa5