Add soft error for cli modifier on ci allowlist via buckconfig
Summary:
Adds a soft error for any cli modifier used in CI that's not in the specified allowlist. This is intended to be used to prevent users from requiring a ton of modifiers on the CLI to be used on the CLI for a target.

More context: https://fb.workplace.com/groups/3454295018049629/permalink/3468426003303197/

Reviewed By: samkevich

Differential Revision: D62106607

fbshipit-source-id: 0c0155e84c1711463126202ebd5816e636ca22a5