Use EdenFS for temp test repo
Summary:
Using EdenFS for our temp e2e test repo. This will be closer to a "real" scenario as most internal users are on Eden.

Note that this will slow tests down by several seconds due to all the additional Eden operations, but it'd only affect Buck2 tests so it should be a fine tradeoff

Reviewed By: stepancheg

Differential Revision: D61457708

fbshipit-source-id: 7a210f322d0dffc3e3f1ac67c7152bac7f02f9b6