Enable setup_eden for a handful of e2e tests
Summary: `Setup_Eden` will cause e2e tests to be slower so enabling it for all tests would be a perf hit. Realistically, we want this to ensure we are running *some* of our tests on Eden (And maybe Watchman for mac later), just in case there's a regression. If that were to happen I think that having 6 tests fail and having 600 tests fail should present the same signal.

Reviewed By: stepancheg

Differential Revision: D62526180

fbshipit-source-id: 6879a46d84f161287a046f76ac089865d928cec5