Add "check" subtarget which typechecks using -fsyntax-only
Summary:
https://releases.llvm.org/18.1.8/tools/clang/docs/ClangCommandLineReference.html#cmdoption-clang-fsyntax-only

Clang's **`-fsyntax-only`** runs the preprocessor, parser and semantic analysis stages. It does *not* run any LLVM IR generation, optimization, assembly, or linking.

The semantics are similar to Rust's `--emit=metadata` which performs typechecking without monomorphization or any LLVM work, except that in C++ template instantiation is required for semantic analysis and in Rust it's not.

Reviewed By: zertosh

Differential Revision: D62622927

fbshipit-source-id: 1ab49cc10317b6d542b592d13136cf1a66d5b5f3