Streaming subprocess stderr
Summary:
Previously, if stderr_to_file.py was killed before arriving to `shutil.copyfileobj(out, sys.stderr)`, we wouldn't see any stderr that the subprocess had printed so far. This doesn't make a difference for a successful `…[check]` subtarget artifact, but does make a difference to debugging when we rerun an action locally and may hit Ctrl+C.

This diff changes to a streaming implementation so that subprocess stderr is copied in real time.

Reviewed By: zertosh

Differential Revision: D62670979

fbshipit-source-id: fd792d88564153313b661cc23745b320ec02e76d