Support fallback config when .pc file isn't available
Summary:
Not all distros (namely ubuntu) publish .pc files for their header packages.
This means that `pkg-config --libs <name>` fails, even though the headers are
avilable.

With this change, if `pkg-config` doesn't recognize the name, it can
potentially fallback to a hardcoded value. It isn't perfect, but gets ubuntu
builds to parity with fedora.

Reviewed By: bigfootjon

Differential Revision: D62789099

fbshipit-source-id: 815ef0f978516c1b5079cf75fad71fdf03f34de2