Do not allocate empty dicts
Summary:
This diff implements a function in `StarlarkValue` for a value to provide statically allocated instance for given mutable value.

During freeze if this function returns `Some`, this value is used, and general freeze is skipped.

This diff uses it to not allocate empty frozen dicts.

Following diff D62780054 uses it to avoid allocation of empty `cmd_args` in buck2.

Reviewed By: JakobDegen

Differential Revision: D62769760

fbshipit-source-id: 4a1fb7c8592b866c7a9a563638f2a9b629e1b2af