completions: Support both zsh installation variants
Summary:
While it is possible to install most zsh completion scripts by sourcing them, the more standard thing to do is instead to place the script onto the fpath, where it will be autoloaded by the completion manager. However, in order for this to work, the first line of the script must be `#compdef buck2`. So adjust the completion template to make that the case.

Also removes the shebangs, which aren't needed for anything

Reviewed By: stepancheg

Differential Revision: D62436680

fbshipit-source-id: 1377ec401ce47ccca37732a945d733b510cec364