completions: Helper binary for e2e tests
Summary:
Our current completion support has some manually written completion scripts specifically for dealing with dynamic completions. We do not use the dynamic completions feature for those, as the clap support is somewhat incomplete.

Even if we didn't have manually written completions though, it would still be good to have full e2e tests of the actual behavior of various shells on the completion scripts we emit.

This diff adds a little helper binary that allows us to test that. The helper binary uses the `complete_pty` crate, which sets up a PTY and isolated shell environment and then goes and runs some shell code. We extract the results to verify the behavior of completions.

I'll plug this into our e2e tests later this stack.

This was super painful to get working and required working around both `complete_pty` limitations and shell bugs. It's fundamentally not the most robust thing ever, but I don't really see us getting around that. Mostly it was bash and zsh that were a pain. Fish, of course, was super straightforward, probably having much to do with it being software from this millenium

Reviewed By: stepancheg

Differential Revision: D62436674

fbshipit-source-id: 6a0c11a646b7bbcdd189a93661f6836322fe76b9