completions: Import fork of `completest_pty`
Summary:
`completest_pty` is a crate for testing shell completions. I was originally using it in the next diff. However, I ended up running into a couple issues with it, a few of which I was able to work around but one of which not. While I could attempt to submit contributions upstream, it looks like this was mostly written for testing `clap_complete` and it's like 500 lines, so easiest to just have a fork of.

This diff imports the one file in the crate verbatim, together with the license files and copyright notices. The file is not associated with a target and intentionally does not build yet, this commit is primarily to keep the contribution history clean

Reviewed By: stepancheg

Differential Revision: D62436678

fbshipit-source-id: ec60b3c62b2f3cfdb7a6d30cf0731c814c3b05e9