Stop attempting to honour pre / post flags
Summary:
Each LinkInfo contains pre_flags, post_flags and a set of linkables. The intended behavior is that the linkables appear on the command line between the pre and post flags. The linkables can be of various kinds (ie. a dynamic library, an object file, a static library).

Darwin has long ignored this. IIRC we list all the flags, then force loaded libraries, then dynamic libraries, then a filelist of everything else.

This diff makes it so that for Darwin dthin-lto builds, we also ignore the pre / post flags semantics.

This is in preparation for D62991973 which orders the linkables according to their type. We can't re-order the linkables from across LinkInfos and reasonably maintain the pre / post flag behavior. Since existing Darwin builds already do this, we do the same.

Reviewed By: rmaz

Differential Revision: D62991971

fbshipit-source-id: a61d3b146e5bc9c7b38832164d3d968c2d14a671