Add tag for windows IO sharing violation errors
Summary:
Errors that look like "The process cannot access the file because it is being used by another process. (os error 32)" on Windows aren't being tagged with anything other than IO_SYSTEM.
It seems rusts stdlib does not assign a kind or provide a kind for these, so do it manually.

This is needed to detect these as environment errors further down.

Differential Revision: D63054608

fbshipit-source-id: 454d6781fcbc3cc496cbdf71790befe546e068dc