Rewrite DocParams
Summary:
Use the same repr as in other places: separate fields for pos-only, pos-or-named, args, named-only, kwargs. Instead of universal params which know what they are.

This repr is
- not redundant
- matches repr in other places in starlark
  - which allows easier conversion
  - and allows reusing signature printing code (which for some reason was implemented twice for `[DocParam]` alone, so this diff reduces the number of printers by two)

This diff also changes `_` with `...` for unknown default/optional for consistency with the rest.

Golden tests changes some functions from multiline to single-line because the logic when to split is not identical.

Reviewed By: JakobDegen

Differential Revision: D63000225

fbshipit-source-id: 5401cca07690377c0ab4f463ef766f69e252cbc2