completions: Enable more fish and zsh tests
Summary:
This diff makes it possible to run some of our tests with fish and zsh on Linux CI. In order to do this, we need to download the associated RPMs at build time.

Specifically, we turn on the unittests on the completion_verify crate for both zsh and fish, and the integration tests for zsh. Integration tests for fish are next.

Unfortunately, I learned a lot about RPMs and zsh doing this. I learned very little about fish, because it works well

Reviewed By: stepancheg

Differential Revision: D63085151

fbshipit-source-id: 7f452f32c2cfec0b739ac9a0986afc836feab3b3