Added new RemoteCacheManagerMode
Summary:
We want to configure Buck's REClient to be able to read from local CAS cache (managed by CASd instance spawned and managed by Turbocache), however, we want to make it **read only**. So, Turbocache should be able to tell CASd to download (**write**) more stuff to the local CAS cache, but Buck should only **read** from it and should not be able to cause the cache to grow. This is important because we want to make sure we only populate local CAS cache with specific targets and not with artifacts for random Buck builds that the user might be performing on their machine.

Following Skory's advice, we are going to:
a) use `RemoteFetchPolicy::LOCAL_FETCH_WITH_SYNC` for both `small_files` and `large_files`.
b) introduce a new enum case `RemoteCacheManagerMode::TURBOCACHE` to not mess with the existing `BIG_FILES` and `ALL_FILES` used by ASIC.

This diff is building on top of D62024564 stack and D63096892 stack.

Reviewed By: ejnguyen

Differential Revision: D63025886

fbshipit-source-id: 96dc53931c42f0cd0b1348e2d21602c77f35acf8