Do not pass parameter name to ParameterParser::next
Summary:
The problem is in code like this:

https://www.internalfb.com/code/fbsource/[92d6211c1a16b8b5708308b538002b562f0a622c]/fbcode/buck2/app/buck2_action_impl/src/dynamic/dynamic_actions_callable.rs?lines=164-169

Parameter names passed to the function are only used for error messages, and there's no check anywhere that this parameter name is correct. So if there's a mistake here, a user gets incorrect error message, which would be hard to catch.

This diff eliminates the parameter, because we already know the parameter name in `ParametersSpec` which `ParameterParser` is created from.

Reviewed By: JakobDegen

Differential Revision: D63292103

fbshipit-source-id: 7c637ff8506d321d69d13871942a20803f94425f