Delete dep files when we know they are no use
Summary:
`check_action` is used to determine if we have an "identical" match on a dep file (i.e. without filtering out the inputs that are not depended on).

If we find that something other than an input has changed, then we return `InitialDepFileLookupResult::Miss` (rather than `InitialDepFileLookupResult::CheckFilteredInputs`). If we get a `Miss` then we should just remove the dep file as we know it is useless.

Reviewed By: scottcao

Differential Revision: D62963991

fbshipit-source-id: c77aa24de7653fbcc0f91b5b41a1d2ef648a3098