Switch Java plugins argument to allow tuples, remove _wip_java_plugin_arguments
Summary:
Earlier, I added `_wip_java_plugin_arguments` as a workaround to allow passing arguments to Java plugins, which wasn't supported. Kotlin plugins are a dictionary `{plugin:arguments}`, while Java plugins are a simple list of plugins. This API is ugly: the separate attribute is not an intuitive design.

Here I am adding support for passing 2-tuples of `(plugin, arguments)` in the existing `plugins` attribute, and removing the `_wip_java_plugin_arguments` attribute. I am consolidating the attribute in the `jvm_common` file because it was repetitive across the different JVM compilation rules.

Reviewed By: IanChilds

Differential Revision: D63460893

fbshipit-source-id: 96b4482061ac7877d0550747e8e404b9f0e51a58