Don't upload to remote dep file cache when it is not enabled but allow_cache_upload=True
Summary:
We call `.cache_upload` in order to upload locally run actions, and to upload dep files.

If we are calling it in order to upload a locally run action, and uploading dep files is not enabled, then we need to make sure we don't pass in our dep file, otherwise it will also be uploaded.

Reviewed By: ezgicicek

Differential Revision: D63532344

fbshipit-source-id: cef7f00c6900c6017d3356300ae40045634f099d