Run typechecker on impl argument of transition() constructor
Summary:
Validate impl function signature at `transition()` call: we fail even if transition is not invoked.

This is not something static typechecker can do, so we run validation at runtime.

Error message is not precise, but better than nothing.

There are reasons for this change:
- while refactoring some bzl files, error comes up earlier. Especially if this transition is not even invoked because of some rare target or something
- error location is better: when transition is invoked with incorrect arguments, it is harder to find the location
- most important one is probably learning. There's an approach to learn API is to insert arbitrary wrong arguments to function calls and get help from the tool what to do next. Now a user may learn what transition is supposed to accept/return even before transition configuration is inserted into some rule, which may be later in the learning path

Finally, another reason for this change was test drive of this machinery before using it in dynamic actions. API of dynamic actions with optional dynamic values is quite complex, and similar machinery should help.

Reviewed By: JakobDegen

Differential Revision: D63504973

fbshipit-source-id: ba802b42be0a8b8c66d2fc1055a31ed2e9e2bb8a