Add resource file duplication check in assemble_bundle
Summary:
Check if the resource files are identical when a conflict is noticed in assemble_bundle, and ignore it when those are the same files.

### Context:
We need to disable `apple_resource_transition` for Messenger Catalyst universal binary build due to some challenges described in this posting. And when it's disabled, we need a way to gracefully handle the conflict that can be resolvable e.g. the files are identical.

https://fb.workplace.com/groups/iosdevxadhoc/permalink/1512627362951871/

Reviewed By: milend

Differential Revision: D63624525

fbshipit-source-id: 8fb00f2b896258f2a77684fa4477aa303ade7c08