Pass the executable name for apple_binary
Summary:
We need a way to control the name of the binary output by the target.

Currently, it defaults to the target's name:

https://www.internalfb.com/code/fbsource/[dbe3e7647eea96f12d57e537105c58de8505250f]/fbcode/buck2/prelude/cxx/cxx_executable.bzl?lines=548%2C808%2C811-815%2C843-844

`executable_name` exists as an attribute on `apple_binary` and when provided, should be used:
https://www.internalfb.com/intern/staticdocs/buck2/docs/prelude/globals/#apple_binary

Set the `executable_name` on `CxxRuleConstructorParams` which is then used in the `cxx_binary_impl` to name the output binary.

Reviewed By: milend

Differential Revision: D63656257

fbshipit-source-id: 932644dbfa9b92a626dc508d88baceb49cf20c23