tests: Start a `tests/core` directory
Summary:
I'll make a brief post about this, but we've talked about this too much at this point and should just do it.

My short term use-case is that I want to diverge the CI configuration between isolated and non-isolated tests.

There are still unresolved issues around how we should be setting up prelude and other non-isolated tests, but we don't have to resolve those in order to start with this.

I'll move a bunch of stuff into here in this stack

Reviewed By: stepancheg

Differential Revision: D63583594

fbshipit-source-id: 4ca7eadfa3f35df6ee7ee53a17f75cc9f4ca1377