Log actions' input files bytes
Summary:
As part of the build speed understability, we want to include additional telemetry about input sizes. In some cases, using a small functionality might pull in a huge dependency causing build speed regressions. For now, this will log the data to build_actions table and will be eventually used to build tooling.

The main changes in this diff:

- New field in `ActionExecutionMetadata`, `RunAction`, and `ExecutedOrReHit` to propagate the input size to `calculation.rs->build_action_no_redirect`

In subsequent diffs:
1.  The field is propagated to the ActionExecutionEnd event via calculation.rs
2. The input_files_bytes is logged to buck2_actions scuba table
3. e2e tests for local and remote execution

Alternatives considered:
1. Make the buck2_actions.rs stateful: We could maintain a map<action_digest_key, input_size> until we see an ActionExecutionEnd event. However this seems very brittle and has the possibility of storing lots of data until we see the actions completing.

Reviewed By: ezgicicek

Differential Revision: D63152185

fbshipit-source-id: b0d4e8c8c9b6cc5da8cf24e2c6d87a8bb87ac2eb