Fix cert check swallowing actual error
Summary:
Currently cert check will swallow majority of the actual error with the exception of top level context. This isn't great as while there are cert issues, it may not actually be the cause and we should display it properly.

There's still a problem where the error is not tagged properly. But that fix will be much easier once the anyhow->buck2_error conversion in buck2/app is complete

Reviewed By: christolliday

Differential Revision: D64009146

fbshipit-source-id: 6d04ebeacf89e17c4758d8276de3c6af3c6a0377