Fix inefficiency in package file computation
Summary:
The package values can't be checked for inequality and so we want to
avoid recomputing them due to irrelevant changes in deps. Add a key for the
package file check (finding the PACKAGE/BUCK_TREE in a directory) so that
we get early cutoff for the packagefile computation itself on irrelevant
changes.

Reviewed By: JakobDegen

Differential Revision: D63668714

fbshipit-source-id: c93fab0c026c86b0d87c78e7ff74412c9d2f1d7d