Hermetic Python toolchain and Buck2 installed packages
The previous setup used Nix to provide a Python toolchain and packages in a reproducible fashion.
However, this requires dedicated remote worker images with the Nix store paths pre-populated which
complicates the setup for testing on Buck2 CI.

Using a hermetic Python toolchain avoids these issues and works on the standard remote execution
image.