Add UnstructuredData instant event
Summary:
There's a lot of buck behavior that is most appropriately tested by
e2e tests and exposing some piece of internal data. Right now, to do that for a
new piece of data requires adding either new top-level subcommands (like audit
commands) or adding new events. That's a little bit heavyweight, so this adds
a single UnstructuredData event that allows us to stuff in anything we want. It's
sort of like the reasoning behind the audit commands implementation (where they
don't require changes to the daemon proto).

See the next diff for an example use.

Reviewed By: stepancheg

Differential Revision: D63722669

fbshipit-source-id: a8c3a2274e39fdc2bf963ca1a99abfe8fcee58fc