Change rendering of "Time elapsed" at command end in expanded progress view
Summary:
This better matches the current behavior, and when we (accidentally)
changed the normal view to leave this on the right side at the end (like the
expanded view was doing) there were complaints about it (i.e. there was at
least some signal that users preferred the current default behavior).

Reviewed By: IanChilds

Differential Revision: D64012708

fbshipit-source-id: 759016f96075d027ad3a22a7304e47f017b9b376