Remove dependency on buck2_configured from buck2_bxl
Summary:
`buck2_bxl` crate is heavy, and having fewer dependencies speeds up incremental compilation.

If this [RFC in Rust](https://github.com/rust-lang/rfcs/pull/3635) will ever be implemented, we will reduce the amount of complexity in code like this.

Initially I wanted to do it to push attribute configuration to downstream crates from `buck2_node` for D63918028, but then decided not to do it for a while. So this diff is not required for D63918028.

Reviewed By: JakobDegen

Differential Revision: D63929039

fbshipit-source-id: 5360c5bb774a4d79471e6dad0ce2c1ba54a6b57b