Refactor batched type check buck rule into standalone function
Summary:
This diff pulls out the type rule to create the batched type checking sub-target into a separate function. In upcoming diffs, shared work will be moved back into the `create_per_target_type_check` function, and then a new function will be created to create each shard type check sub-target.

 ---

I'm going to be making type checking operations shardable, which should speed up type checking (especially for IDE use cases). The method of doing this will be to create sub-targets under the `typecheck` sub-target, all of which we will send to RE to calculate type errors.

The idea here is to kick off separate RE jobs for each file, where we type check only that file. All other files in the target (and all dependencies) are treated as dependencies, and are not type checked.

Followup: type check only the shards relating to the specific files that are open (for IDE use).

effective_sharding

Reviewed By: grievejia

Differential Revision: D64063902

fbshipit-source-id: db57b7b7ba70131c7b3341c82954b8d874eabc4f