Add RE Reason group to RE error category key
Summary:
Initially attempted as D63058101

Having separate ErrorTags for each RE reason group is going to be painful to keep updated, instead let's just handle them as strings so that they are always updated.

This uses `context_for_key` to append the reason group to the end of the category key, which means it's not actually a tag but can still be used to group errors. Support for having separate (dynamic) tags for this can be added later.

Reviewed By: Will-MingLun-Li

Differential Revision: D63730990

fbshipit-source-id: 077132b9703f3e79ea748792efd085fe1be7f8e6