Replace downcasting of RemoteExecutionError
Summary:
Downcasting anyhow to RemoteExecutionError is a problem for 2 reasons:
- I want to convert RemoteExecutionError to `buck2_error::Error` in the following diff to attach some context that can't be added to a non-buck2_error
- It won't work once we replace `anyhow::Error` with `buck2_error::Error` for buck2 internal errors

Instead, this stores a copy of the RemoteExecutionError as TypedContext and uses `find_typed_context` to find and downcast it.

In theory we could use `has_tag` to determine the `TCode` and skip downcasting entirely, however there is code that should check `TCodeReasonGroup` (but currently checks the error message), and it's preferable not to create tags for every `TCodeReasonGroup`.

It's a bit wasteful to store a copy of the original error as context, but this should ultimately be cleaned up to not string match RE error messages and to only store `TCode` and `TCodeReasonGroup`.

Reviewed By: Will-MingLun-Li

Differential Revision: D63872101

fbshipit-source-id: dfa067078efb34386dd842c649a6e7dec46794d3