Multiple header units per lib
Summary:
This provides some flexibility to modularise existing libraries in a more fine-grained fashion. Ideally, libraries should be buckified along the lines of the modules we want to generate (one header unit per target), but this is not always readily feasible. In order, to aid with migration and performance, we introduce the ability to partition the headers into different header units.

`export_header_unit_filter` now accepts a list of regexes, each defining a different group of headers to turn into a unit. Headers will be intercepted and load the corresponding header unit. The list is built sequentially so that headers in subsequent groups can depend on headers in earlier groups.

Reviewed By: apolloww

Differential Revision: D64062362

fbshipit-source-id: e6493ec9e34fbca28a3184f32c05d469c5e40705