Add invalidation tracking
Summary:
Dice computations track "invalidation paths" for every node in the computation.

Invalidation path tracking is a data flow analysis that tracks the flow of changed
state through the computation graph. Directly invalidated or injected nodes
are the root "invalidation sources". For every computed node, its invalidation paths
will each either be "clean" (if all of its dependencies' invalidation paths are clean)
or will point to one of its non-clean dependencies . When multiple deps
are not clean, it will point to the one with the most recently changed invalidation
source. In this way, we record an invalidation path through the graph to the invalidation
source for every affected node.

We track invalidations of two priorities: "normal" and "high". Each key type specifies
its own invalidation source priority that is used when that node is directly invalidated or
injected. Each node will record the most recent invalidation source for each priority.
A high priority invalidation will also count for normal priority tracking (but not the
reverse). So if a node has two deps A and B, and A has a normal priority invalidation
at v1 and B has a high priority invalidation at v2, B would be recorded for both normal
and high priority source. If those versions were reversed (A at v2, B at v1), we'd
record A for the normal priority invalidation path (because its invalidation is more recent) and would
record B for the high priority invalidation path (because A is normal priority).

Context:

The primary goal here is to be able to know during any action execution whether
it has been tainted by a file change. This in turn is so that we can explicitly track
two things:

1. for each action execution, was it affected by a user's changes
2. for each action error, was it affected by a user's changes

(1) is really about being able to ensure that cache warming is working as expected at
scale. (2) has the potential to allow us to be able to have high signal that a broken
diff has landed.

Reviewed By: stepancheg

Differential Revision: D60150441

fbshipit-source-id: b7515e3ad4dd1a359ca1424ebad6dcc53759dc67