Add taint information in ActionExecutionEnd
Summary:
records the taints from dice into this event. This will allow us to
add this information to logging to support both:
(1) understanding when an executed action was or wasn't affected by a user's
changes (which will help identify cases where cache warming isn't working)
(2) understanding when an action failure was or wasn't affected by a user's
changes (which should allow us to get better signal for trunk breakages).

Reviewed By: stepancheg

Differential Revision: D60534626

fbshipit-source-id: 7561a39d20c7e7af5e840944b89bc3a89318b341