We should not write into projected artifact
Summary:
We can do `a.project("xxx").as_output()`, and write to it, and it is successful.

And on top of that, we can read what is "written" into `a`.

It does not make sense, I think we should not allow using projected artifacts as output artifacts. I.e. `a.project("xxx").as_output()` should fail. Even if we don't attempt to write to it. Right?

Reviewed By: JakobDegen

Differential Revision: D64194174

fbshipit-source-id: 2902613ba271060c10305d0c39e980a07142babc