log cesid
Summary:
#### adding CESID to buck2_builds

edited the InvocationRecord (which accroding to [data.proto](https://fburl.com/code/mnpbp0go) its the datastruct logged to scuba)

and added to ActionExecutionEnd for debugging purposes

### Why
buck team wants to run experiments on the new `eagerbuild` feature.  given the prioritization of dev-environments over users as population ID for the experiment, we need to first log the [Cross Environment Session Id](https://fb.workplace.com/groups/301619822315138/posts/502477458896039) for every buck build, to then be able to build a data pipeline based on the info provided on buck2_builds scuba table.

this is all a necessary first step to get experimentation data based on development environments.

Reviewed By: JakobDegen

Differential Revision: D64130138

fbshipit-source-id: d68e9876f7196c125217e0d57fc15bf49715161e