Remove ProjectedArtifact
Summary:
We treat regular artifacts and projected artifacts the same everywhere. Also we imply that projected artifact path is not empty.

For example, `artifact.project("")` is a valid operation, which just returns self.

So instead of thinking we have different artifact kinds, this diff changes artifact representation as always base artifact and projection path, where the path can be empty.

The change is from:

https://www.internalfb.com/code/fbsource/[4824c5972bf9c93cb438a7ca057087b3ee23468a]/fbcode/buck2/app/buck2_artifact/src/artifact/artifact_type.rs?lines=203-206

to

https://www.internalfb.com/code/fbsource/[D64224117-V5]/fbcode/buck2/app/buck2_artifact/src/artifact/artifact_type.rs?lines=198-202

Reviewed By: JakobDegen

Differential Revision: D64224117

fbshipit-source-id: bd6c2952a42ddf6945875e7d83b84950675c409a