Move RFC to google doc
Summary:
This RFC is what people are referencing but is constantly getting out-of-date due to modifier API changes.

Move the RFC itself (the API page) to a google doc so it stays more up-to-date and enables comments.
Deleted the use cases page because it's difficult to keep this up-to-date, and it's much simpler to add back a page of real examples of modifiers instead.

For staticdocs, I changed existing RFC page to be internal-only and point to the google doc.

Unfortunately, this does mean that OSS no longer gets to see the RFC, but I'm fixing that in next diff by publishing it as a PDF instead.

Reviewed By: Will-MingLun-Li

Differential Revision: D64274616

fbshipit-source-id: 3eee097c7fc728a68464a60324b72655336cb738