tests: Fix `setup_eden`
Summary:
This wasn't using the actual project dir as the eden mount point. Fix that.

This bug was at least partially the result of some bad variable naming, so fix that too.

This revealed a new issue, where the relevant tests were failing with soft errors due to a missing buck-out redirect. So add in a redirect.

Reviewed By: IanChilds

Differential Revision: D64294153

fbshipit-source-id: bcf2be56dcdcdfa921c07cb1d2eedd9b1fd25f00