Refactor derive(Freeze)
Summary:
Use the same parser for `#[freeze]` on struct and on field. Errors with better spans.

For future changes. What we need is something like

```
#[freeze(frozen = ...)]
```

to `derive(Freeze)` when some fields are generic but freeze is identity. Currently it is not possible, because even if bounds is specified, generated code still uses `T::Frozen` for frozen type parameter.

Reviewed By: JakobDegen

Differential Revision: D64337850

fbshipit-source-id: 970c3ade11061b9f3bd4f55baabf78cc076a9f6d