Only flush non-local dep files when the mergebase changes
Summary:
The comment explains the reasoning behind the current behavior. For locally produced dep files, the trade off is different - they are not cached anywhere else, and we already have the dep file on disk, so it makes a lot more sense to retain them when we can.

This diff makes it so that we retain them after a rebase (or change of "mergebase").

Reviewed By: milend

Differential Revision: D63318010

fbshipit-source-id: cbba92f0161473bee407b72856858310694ee7f4