Add was_produced_locally to DepFileState
Summary:
We want to start retaining locally produced dep files in more scenarios (because locally run actions are not cached elsewhere, and also we don't have to download the dep-file, we already have it locally).

So, to begin with we need to keep track of whether the dep file was produced locally or not.

Reviewed By: scottcao

Differential Revision: D63318012

fbshipit-source-id: cfec7ba5dc96020176f3c1e83bc5d61e0e538829