Use `scribe_cat` to log to scuba
Summary:
Using `scuba` directly goes through the scribed thrift client. That thing is slow, both for binary startup time and when running it.

Also, this version has the advantage that we can detach the scribe_cat process. That means if something there is slow, we won't wait on it.

None of this matters much for `rust-project develop`, but it definitely matters for `rust-project check`.

Reviewed By: perehonchuk

Differential Revision: D64361425

fbshipit-source-id: a3a8aa0319260aeec3785dca46673b572cb3488d