Only log actions stats for run actions
Summary:
When we log action stats to InvocationRecorder, we only do that for actions with Run kind.

https://www.internalfb.com/code/fbsource/[29c2795794e2]/fbcode/buck2/app/buck2_client_ctx/src/subscribers/recorder.rs?lines=870-881

However, when we print action stats at the end of the build, this check is missing which seems wrong.

I uncovered this issue when I was trying to display skipped_action count in the next diff where despite buck UI saying 0 skipped acitons, I got 100K skipped actions.

Reviewed By: iguridi

Differential Revision: D63393753

fbshipit-source-id: 3fcbafd675cb0f2c333aebfe6aacdfc8d8ec37cd