Use RE error as source for CAS materialization errors
Summary:
We aren't currently tagging RE errors that cause materialization failures because the tags for the underlying RE error are thrown away.

Using the RE error as the source for `MaterializationError::NotFound` means maintaining all RE error tags, including string context (TCodeReasonGroup) used as tags in the category key.

This changes the error shown to the user, since we are adding a second piece of error context on top of an underlying RE error, from this:
```
Failed to materialize artifacts

Caused by:
    Your build requires materializing an artifact that has expired in the RE CAS and Buck does not have it. This likely happened because your Buck daemon has been online for a long time. This error is currently unrecoverable. To proceed, you should restart Buck using `buck2 killall`.

    Debug information:
      Path: buck-out/v2/gen/fbsource/b61af97434061b22/fbandroid/java/com/facebook/tools/editus/__editus__/jar
      Digest origin: 23362b641d9852e58ddc8ac3e3d503f294391ad53cbd437119b486816141be71:145 retrieved 0 seconds ago with ttl = 77880 seconds (not expired: action cache corruption)
      Debug info: Injected error
      Directory: ..
```
to this:
```
Failed to materialize artifacts

Caused by:
    0: Your build requires materializing an artifact that has expired in the RE CAS and Buck does not have it. This likely happened because your Buck daemon has been online for a long time. This error is currently unrecoverable. To proceed, you should restart Buck using `buck2 killall`.

       Debug information:
         Path: buck-out/v2/gen/fbsource/b61af97434061b22/fbandroid/java/com/facebook/tools/editus/__editus__/jar
         Digest origin: 23362b641d9852e58ddc8ac3e3d503f294391ad53cbd437119b486816141be71:145 retrieved 0 seconds ago with ttl = 73487 seconds (not expired: action cache corruption)
         Directory: ..

    1: Remote Execution Error on test for ReSession test
       Error: (Injected error)
```

Arguably the older error is a bit more readable, however matching that while copying the tags correctly from the underlying error is not simple.
Either we would need to add a custom `std::error::Error`/`provide` impl  for MaterializationError or change the type where it is returned to `buck2_error::Error` and lose the ability to match unless we get typed context and downcast.
Neither of these seem worth it considering this is an infra error that users typically shouldn't see, except when buck is automatically restarting, and the error with the added context seems potentially more informative and not much longer.

Reviewed By: IanChilds

Differential Revision: D64353582

fbshipit-source-id: 8b84c36d1b5845a08d229e32c93c1468846d5cf4