Enable sharding with batch_files.bxl query
Summary:
This diff adds the argument to shard type checks to batch_files.bxl.
 ---

I'm going to be making type checking operations shardable, which should speed up type checking (especially for IDE use cases). The method of doing this will be to create sub-targets under the `typecheck` sub-target, all of which we will send to RE to calculate type errors.

The idea here is to kick off separate RE jobs for each file, where we type check only that file. All other files in the target (and all dependencies) are treated as dependencies, and are not type checked.

Followup: type check only the shards relating to the specific files that are open (for IDE use).

effective_sharding

Reviewed By: grievejia

Differential Revision: D64063892

fbshipit-source-id: c64c3cc525b98d212eae8cdfd8b70345d5aaf49c