Change cold build detector errors to soft errors
Summary: The cold build detector was added in D63633614 to assist with cache miss warning. However, this has been causing crashes when the merge base changes. This was originally expected not to change but it does show up when the file watcher events are generated for a 'Fresh instance' or when the count of file changes are high.

Reviewed By: samkevich

Differential Revision: D64465953

fbshipit-source-id: 2f0156f17f17601c9d26e0d84e1d393a4c1ff084