Default to host sys platform in PTT
Summary:
just_a_platform

We've already got `py_version_for_type_checking` to help per-target Pyre to override the python version number so it can do a better job at resolving conditional branches like `if sys.version >= (X, Y)`. We've recently added similar logic to better resolve branches like `if sys.platform == XYZ` (D63336519), but what exactly is the value of `sys.platform` is something that has to be determined externally from the type checker. In the case of Buck-built projects, we'd want the info to come from Starlark logic.

This diff adds some basic heuristic on guessing what `sys.platform` is from the Buck side, and pass that down to the Pyre binary. This should work in most cases but probably won't work if we are cross-compiling / cross-building. If the need came up later, we might re-consider D64129800(V2) to make it possible to override the platform by individual targets.

Reviewed By: zsol

Differential Revision: D64129799

fbshipit-source-id: f8f8d5572abfe53c8fff53e495928f668bbf1aa8