Fix nested validation of expressions
Summary:
We have a validation pass that looks for things like `lambda a,a:a` and turns it into an error. Previously this looked for the top-level expression directly inside a statement. That means it missed things like `lambda: lambda a,a:a`. Easily fixed by just recursing in the visit_expr call. Our validation is assumed to have happened by the compiler, so that code would lead to:

```
assertion failed: old_local.is_none()
```

Caught by the OSS Fuzzer that we have set up.

Reviewed By: IanChilds

Differential Revision: D64663606

fbshipit-source-id: 13c71abb4609be75f8879dfdf9eb2706c53d6dd1