Support test_deps that aren't foo-unittest
Summary:
Previously, we assumed that any test target that occurred in `test_deps` was a generated unittest that had a name of the form `foo-unittest`. This isn't always true, and led to us discarding test dependencies that had other names.

Instead, be stricter about target names, so other test dependencies end up in `standalone_tests` as expected.

Reviewed By: darichey

Differential Revision: D64634915

fbshipit-source-id: 4cdbd2a766baa9515b13c68c415a141d4cf0efe6