make sidecar a layer for __unstamped jars
Summary:
The fat-jar process does merge all dependencies by opening each entry and deflating it again. Thats specially expensive when using sidecar, given the content can be up to 12GB of shared libs.

The idea here is to use the sidecar, as a base jar, and increment a copy of the jar instead of repacking it from scrach at each stage of the jar generation.

 {F1920027649}

Once sidecar_lib_cosco-* is cached, all subsequent builds will only append to a copy instead of zip at each build

Reviewed By: IanChilds

Differential Revision: D64146985

fbshipit-source-id: 22898c599117bc6337ac3a19806f84ec0405bd63