Write error_category to invocation record
Summary:
Currently error_category written to buck2_builds is determined by ingress based on the errors in the invocation record.

I want to move this logic into buck2 in order to make it more predictable and easier to test and to change how it's computed. Specifically I want to make it so that the category reported is the same as the category of best_error_tag and best_error_category.

This just copies the logic from ingress 1:1, will change how it is computed later.

Reviewed By: JakobDegen

Differential Revision: D64561177

fbshipit-source-id: b4c520c5d87307f59d3b5f7662fc6933c370afd5