Use error tier for CommandExecutionError
Summary:
`CommandExecutionError`s that aren't "command failures" will only be marked as input errors if they have a tag that is classified as an input tag, not if it's only marked `#[buck2(input)]`.

Fix it by re-using the category/tier.

I think it would be better to exclusively use tags to derive categories, but for now let's just make this work.

Reviewed By: Will-MingLun-Li

Differential Revision: D64572362

fbshipit-source-id: 1df60106b312ff094713f0a1a57962543f60d574