Make postprocessed jars deterministic
Summary:
Let's make post processed jars deterministic. Currently, datetime is included in the jar output when you run the command in the test plan. This prevents caching so let's scrub the jar once we are done post processing it.

The data and time will stay constant now.

References:
https://www.internalfb.com/diff/D50930818
https://www.internalfb.com/diff/D38073001
https://www.internalfb.com/diff/D57650082

This diff was previously reverted in https://www.internalfb.com/diff/D63637702.

Reviewed By: navidqar

Differential Revision: D63857547

fbshipit-source-id: c91ae0107a1194a14f538e4c1ebe3f1ee32c3f29