Add starlark->buck2_error callstack test
Summary:
Adding a test for starlark->buck2_error conversion where starlark has a call stack.

Tests ensures that converting to buck2_error would preserve the call stack and will print it on top of the error stack no matter what (Otherwise it will look weird visually). Also ensures that if there are multiple call stacks (Not sure how practical it is), all of them will be displayed in the same fashion

Reviewed By: christolliday

Differential Revision: D64092927

fbshipit-source-id: cf95e5ea7326c7f66ce242418819f69cf33426d0