Updating the force_tests_on_mac_re mode file to only contain the RE use case informtion
Summary:
With the goal of having less mode files in arvr. The mode file to force mac tests to run on RE will eventually be removed as this diff sets up the toolchain to ensure that tests can run on RE if there is a platform mismatch D62482846.

Currently the use case is being used to track the increase in mac capacity while jobs are being migrated over from mac-arm to linux. Once the migration has been completed the mode file will be deleted, and the use case RE capacity for `rl-mac-default` will be merged with that of `re_tests`

Reviewed By: 8Keep

Differential Revision: D64796922

fbshipit-source-id: c52bd3176edd3fe70f526afa369daab9df786615