Removing platform information from the force_tests_on_mac_re modefile
Summary:
With the goal of having less mode files in arvr. The mode file to force mac tests to run on RE will eventually be removed as this diff sets up the toolchain to ensure that tests can run on RE if there is a platform mismatch D62482846. Now the modefile will only contain information to indicate whether to force the tests to run on mac RE (regardless of what platform it is run from)

Currently the use case is being used to track the increase in mac capacity while jobs are being migrated over from mac-arm to linux. Once the migration has been completed the mode file will be deleted, and the use case RE capacity for rl-mac-default will be merged with that of re_tests

Reviewed By: 8Keep

Differential Revision: D64994277

fbshipit-source-id: d084b400b023ef1cd2578badb857041560eb1831