docs: Switch to mdx everywhere
Summary:
First, some context. Docusaurus3 supports two markdown rendering strategies. The first is commonmark which is just bare-bones traditional markdown rendering without support for anything fancy. The second is mdx, which allow embedding react components in the markdown.

The docusaurus2 -> docusaurus3 upgrade accidentally broke some things on the doc site. The result of this breakage is that docusaurus2 was using weird heuristics to try and detect when a page should be treated as commonmark vs mdx, and was also using an older version of mdx. Docusaurus3 instead takes the approach of having you specify which you want, either globally or by file extension. In the upgrade diff I had specified commonmark, but it turns out we were relying on mdx for a small number of things, and those things then broke. Reported here: https://fb.workplace.com/groups/buck2users/permalink/3792154464374240/

This diff fixes thing by switching us to mdx everywhere. That requires adjusting some things to deal with newer versions of mdx, primarily that an import is now needed for the `<FbInternalOnly>` react component.

I had instead considered moving just the pages that needed it to mdx, however rejected this because I have reason to believe we'll eventually want this anyway (and also it wouldn't have actually been easier to do that)

Reviewed By: blackm00n

Differential Revision: D65005718

fbshipit-source-id: eb37a2eb4accd8289673fd66500b2cca452b8734