Reland [buck2] docs: Typescript
Summary:
Reland of D61778034 , with D64857942 merged in.

These were reverted in D64896964

The main change is that I renamed the `.ts` config file from `docusuaurs.config.ts` to `config_impl.ts`. The part where there was both a `.js` and a `.ts` of the same name was breaking things in a way that didn't show up in CI

Reviewed By: blackm00n

Differential Revision: D65006831

fbshipit-source-id: 1654bc93e98b5fcabb184246e16969655824e864