docs: Add remaining redirects
Summary:
To match the page name changes in D61778036

I diff'd the generated docs before and after to find the pages that had disappeared and added redirects for all of them. In some cases, this was not possible as the docs for those types are not generated anymore - this happened specifically for types that we were documenting but don't have a type symbol for.

Reviewed By: blackm00n

Differential Revision: D65018405

fbshipit-source-id: ce150e44e00c667603c5d7ca4518135e91edd3ca